==> 2-clean.ipynb <==
# Notebook 2:  Data Cleaning 

In this notebook we merge our raw data tables from Notebook 1 and clean the player salaries.
1.  **Merging**: We update our 2020 free agent data with the 2021 next-season salaries, and also merge current-season (2021-22) stats and salaries together.  This gives us two new dataframes:  df2020_merged and df2021_merged.
2.  **Cleaning salaries**: We convert salaries into numerical form.  We also normalize salaries from all years to the 2021 salary cap for consistency.

## Imports

%matplotlib inline
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
import pandas as pd

from bs4 import BeautifulSoup, Comment
import requests
import lxml
import unicodedata
import json, pickle

## Utility Functions

def strip_accents_and_punctuation(text):
    '''Normalize player name spellings'''
    try:
        text = unicode(text, 'utf-8')
    except NameError: # unicode is a default on python 3 
        pass
    text = unicodedata.normalize('NFD', text)\
           .encode('ascii', 'ignore')\
           .decode("utf-8")
    return str(text).replace('.','').replace(',','').replace("'",'')

# (1) Merge Scraped Data

df2020 = pd.read_csv('./data/df2020_raw.csv')
df2021salaries = pd.read_csv('./data/dfcurrentsalaries.csv')
df2021stats = pd.read_csv('./data/dfcurrentstats.csv')
df2021urls = pd.read_csv('./data/dfplayer_to_url.csv')

## a. Update df2020 with 2021 salaries

df2020_merged = pd.merge(df2020, df2021salaries, left_on="Name", right_on="Name", how="left")
df2020_merged = df2020_merged.drop(columns=['NextSal', 'Unnamed: 0', 'CurrentTeam'])
df2020_merged = df2020_merged.rename(columns = {'CurrentSalary':'NextSal'})
df2020_merged = df2020_merged.set_index('NameYear')
df2020_merged.head(3)

## b. Update df2021 with 2021 salaries and player urls

df2021_merged = pd.merge(df2021stats, df2021salaries, left_on="Player", right_on="Name", how="left")
df2021_merged = pd.merge(df2021_merged, df2021urls, left_on="Player", right_on="Name", how="left")

df2021_merged = df2021_merged.drop(columns=['Name_x', 'Name_y','Unnamed: 0','Unnamed: 0_x','Unnamed: 0_y','Rk'])
df2021_merged = df2021_merged.rename(columns = {'Player':'Name'})

df2021_merged.head(3)

## c. Update df2021 with current team and total stats

This is a little bit tricky because some players played for multiple teams.  As an example, Nickeil Alexander-Walker switched teams mid-season and thus has three data rows:

* Name    /   Team
* Nickeil Alexander-Walker  /    TOT
* Nickeil Alexander-Walker  /    NOP
* Nickeil Alexander-Walker  /    UTA

We want to keep his **season stats** from the first row (TOT) but also indicate his **current team** from the last row (UTA).  The same pattern holds for all players who played for multiple teams.

# Create current team name filter 
justteam = ~df2021_merged.duplicated(subset=['Name'],keep='last')
# Create total stats filter
juststats = ~df2021_merged.duplicated(subset=['Name'],keep='first')


df2021_merged = (pd.merge(df2021_merged[juststats], df2021_merged[justteam][['Name','Tm']], 
                          left_on="Name", right_on="Name", how="left")
                            .rename(columns ={'Tm_x':'TeamOrTot', 'Tm_y':'Team'}) 
         )

df2021_merged['NameYear'] = df2021_merged['Name'].apply(strip_accents_and_punctuation).str.replace(' ','_') + '_2021' 
df2021_merged = df2021_merged.set_index('NameYear')

Minor bookkeeping:
Update a few Teams with correct info (only applies to 15 players, not that important):


mask = df2021_merged.CurrentTeam.notna() & (df2021_merged.CurrentTeam != df2021_merged.Team)
df2021_merged.loc[mask,'Team'] = df2021_merged.loc[mask,'CurrentTeam']

df2021_merged = df2021_merged.drop(columns=["CurrentTeam"])
df2021_merged = df2021_merged.rename(columns= {'CurrentSalary': 'Salary'})

df2021_merged.head(5)

## Save merged data

df2020_merged.to_csv(r'/Users/andrei/Dropbox/Metis/HoopsHero/data/df2020_merged.csv')
df2021_merged.to_csv(r'/Users/andrei/Dropbox/Metis/HoopsHero/data/df2021_merged.csv')

# (2) Clean salary data

## Load files

df2016_raw = pd.read_csv('./data/df2016_raw.csv')
df2017_raw = pd.read_csv('./data/df2017_raw.csv')
df2018_raw = pd.read_csv('./data/df2018_raw.csv')
df2019_raw = pd.read_csv('./data/df2019_raw.csv')
df2020_merged = pd.read_csv('./data/df2020_merged.csv')

df  = pd.concat([df2016_raw, df2017_raw, df2018_raw, df2019_raw, df2020_merged])
df = df.set_index('NameYear')

df2021 = pd.read_csv('./data/df2021_merged.csv')
df2021 = df2021.set_index('NameYear')
df2021 = df2021.drop("Player_2021")  #spurious entry

## Delete unnecessary entries

We first employ a very light common-sense filter on our entries.  To train our player value model, we wish to keep only players who: 
1.  Played in at least one game the previous year
2.  Had a salary the next year (i.e. actually received a contract in free agency)

#  1. Mask for players who played in at least one game the previous year

df['G'] = df['G'].apply(lambda s: pd.to_numeric(s, errors='coerce'))
mask1 = df['G'].isna()
df = df[~mask1]

#  2. Mask for players who had a salary the next year

mask2  = (df['NextSal'] == '0') | df['NextSal'].isna()  
df = df[~mask2]
df.sample(3)

### Clean salries and normalize by salary cap

Since we are looking at data for multiple years, it is important to normalize by salary cap for apples-to-apples comparisons in our model.  We chose the 2021 salary cap for simplicity.

def clean_salary(sal, year):
    ''' Adds all salaries earned per year into one salary (in millions of dollars)
        "< $Minimum" is treated as $0
    '''
    newlist = [0]
    for elem in sal.split('$'):
        try:
            elem = elem.replace('<','').replace(',','').replace('(TW)','').replace('Minimum','0')
            elem =  float(elem)
            newlist.append(elem)
        except:
            continue
            
    salary = sum(newlist)/1000000
    
    return salary

# Salary Cap Data taken from spotrac.com/nba/cba
salarycap = {2016:   94.143,  2017: 99.093,   2018: 101.869,   2019:  109.14,   
             2020:   109.14,  2021: 112.414}   

def clean_and_norm_salary(sal, year):
    ''' Adds all salaries earned per year into one salary (in millions of dollars)
        "< $Minimum" is treated as $0
        Then normalizes by salary cap (in 2021 salary cap dollars)
    '''
    newlist = [0]
    for elem in sal.split('$'):
        try:
            elem = elem.replace('<','').replace(',','').replace('Minimum','0')
            elem =  float(elem)
            newlist.append(elem)
        except:
            continue
            
    salary = sum(newlist)/1000000
    salary = salary*salarycap[2021]/salarycap[year]
        
    return salary

for index, row in df.iterrows():
    
    prevyear = df.loc[index, 'PrevYear']
    nextyear = df.loc[index, 'PrevYear'] + 1 

    prevsal = df.loc[index, 'PrevSal']
    nextsal = df.loc[index, 'NextSal']
    
    df.loc[index, 'PrevSalClean'] = clean_salary(prevsal, prevyear)
    df.loc[index, 'NextSalClean'] = clean_salary(nextsal, nextyear)
    df.loc[index, 'PrevSalNorm'] = clean_and_norm_salary(prevsal, prevyear)
    df.loc[index, 'NextSalNorm'] = clean_and_norm_salary(nextsal, nextyear)

cols = ['NextSal', 'NextSalClean','NextSalNorm']
df[cols].sample(8)

for index, row in df2021.iterrows():
    
    salary = df2021.loc[index, 'Salary']
    try:
        df2021.loc[index, 'SalClean'] = clean_salary(salary, 2021)
    except:
        df2021.loc[index, 'SalClean'] = np.nan
        
    try:
        df2021.loc[index, 'SalNorm'] = norm_salary(salary, 2021)
    except:
        df2021.loc[index, 'SalNorm'] = np.nan

df2021.sample(3)

## Save cleaned data

df.to_csv(r'/Users/andrei/Dropbox/Metis/HoopsHero/data/cleaned_pastyears.csv')
df2021.to_csv(r'/Users/andrei/Dropbox/Metis/HoopsHero/data/cleaned_currentyear.csv')
